fix: migrate dbconnection
