+++ src/app.ts
@@ -26,6 +26,7 @@ import { keys } from "./routes/management-api/keys";
 import { tailwindCss } from "./styles/tailwind";
 import authenticationMiddleware from "./middlewares/authentication";
 import { logoutRoutes } from "./routes/oauth2/logout";
+import { dbConnectionRoutes } from "./routes/tsoa/dbconnections";
 
 const ALLOWED_ORIGINS = [
   "http://localhost:3000",
@@ -93,6 +94,7 @@ export const loginApp = rootApp
   .route("/u", login)
   .route("/.well-known", wellKnown)
   .route("/callback", callback)
+  .route("/dbconnections", dbConnectionRoutes)
   .route("/v2/logout", logoutRoutes)
   .route("/api/v2/domains", domains)
   .route("/api/v2/users", users)
