+++ test/integration/flows/password.spec.ts
@@ -30,24 +30,16 @@ describe("password-flow", () => {
   describe("Register password", () => {
     it("should return a 400 if an invalid client is passed", async () => {
       const env = await getEnv();
-      const client = testClient(tsoaApp, env);
+      const client = testClient(loginApp, env);
 
-      const typesDoNotWorkWithThisSetup___PARAMS = {
+      const response = await client.dbconnections.signup.$post({
         json: {
           client_id: "invalidClientId",
           connection: "Username-Password-Authentication",
           email: "test@example.com",
           password: "Password1234!",
         },
-      };
-      const response = await client.dbconnections.signup.$post(
-        typesDoNotWorkWithThisSetup___PARAMS,
-        {
-          headers: {
-            "content-type": "application/json",
-          },
-        },
-      );
+      });
       expect(await response.text()).toBe("Client not found");
 
       expect(response.status).toBe(404);
@@ -57,23 +49,15 @@ describe("password-flow", () => {
       const password = "Password1234!";
       const env = await getEnv();
       const client = testClient(tsoaApp, env);
-
-      const typesDoNotWorkWithThisSetup___PARAMS = {
+      const loginClient = testClient(loginApp, env);
+      const createUserResponse = await loginClient.dbconnections.signup.$post({
         json: {
           client_id: "clientId",
           connection: "Username-Password-Authentication",
           email: "password-login-test@example.com",
           password,
         },
-      };
-      const createUserResponse = await client.dbconnections.signup.$post(
-        typesDoNotWorkWithThisSetup___PARAMS,
-        {
-          headers: {
-            "content-type": "application/json",
-          },
-        },
-      );
+      });
       expect(createUserResponse.status).toBe(200);
 
       const loginResponse = await client.co.authenticate.$post(
@@ -244,22 +228,14 @@ describe("password-flow", () => {
         login_count: 0,
       });
 
-      const typesDoNotWorkWithThisSetup___PARAMS = {
+      const createUserResponse = await loginClient.dbconnections.signup.$post({
         json: {
           client_id: "clientId",
           connection: "Username-Password-Authentication",
           email: "existing-code-user@example.com",
           password,
         },
-      };
-      const createUserResponse = await client.dbconnections.signup.$post(
-        typesDoNotWorkWithThisSetup___PARAMS,
-        {
-          headers: {
-            "content-type": "application/json",
-          },
-        },
-      );
+      });
       expect(createUserResponse.status).toBe(200);
 
       // -----------------------------
@@ -405,23 +381,16 @@ describe("password-flow", () => {
       const password = "Password1234!";
       const env = await getEnv();
       const client = testClient(tsoaApp, env);
+      const loginClient = testClient(loginApp, env);
 
-      const typesDoNotWorkWithThisSetup___PARAMS = {
+      const createUserResponse = await loginClient.dbconnections.signup.$post({
         json: {
           client_id: "clientId",
           connection: "Username-Password-Authentication",
           email: "password-login-test@example.com",
           password,
         },
-      };
-      const createUserResponse = await client.dbconnections.signup.$post(
-        typesDoNotWorkWithThisSetup___PARAMS,
-        {
-          headers: {
-            "content-type": "application/json",
-          },
-        },
-      );
+      });
       expect(createUserResponse.status).toBe(200);
 
       const loginResponse = await client.co.authenticate.$post(
@@ -526,9 +495,10 @@ describe("password-flow", () => {
     it("should not allow a new sign up to overwrite the password of an existing signup", async () => {
       const env = await getEnv();
       const client = testClient(tsoaApp, env);
+      const loginClient = testClient(loginApp, env);
       const aNewPassword = "A-new-valid-password-1234!";
 
-      const typesDoNotWorkWithThisSetup___PARAMS = {
+      const createUserResponse = await loginClient.dbconnections.signup.$post({
         json: {
           client_id: "clientId",
           connection: "Username-Password-Authentication",
@@ -536,15 +506,7 @@ describe("password-flow", () => {
           email: "foo@example.com",
           password: aNewPassword,
         },
-      };
-      const createUserResponse = await client.dbconnections.signup.$post(
-        typesDoNotWorkWithThisSetup___PARAMS,
-        {
-          headers: {
-            "content-type": "application/json",
-          },
-        },
-      );
+      });
 
       expect(createUserResponse.status).toBe(400);
       const body = await createUserResponse.text();
@@ -570,24 +532,16 @@ describe("password-flow", () => {
     });
     it("should reject signups for weak passwords", async () => {
       const env = await getEnv();
-      const client = testClient(tsoaApp, env);
+      const client = testClient(loginApp, env);
 
-      const typesDoNotWorkWithThisSetup___PARAMS = {
+      const createUserResponse = await client.dbconnections.signup.$post({
         json: {
           client_id: "clientId",
           connection: "Username-Password-Authentication",
           email: "weak-password@example.com",
           password: "password",
         },
-      };
-      const createUserResponse = await client.dbconnections.signup.$post(
-        typesDoNotWorkWithThisSetup___PARAMS,
-        {
-          headers: {
-            "content-type": "application/json",
-          },
-        },
-      );
+      });
 
       expect(createUserResponse.status).toBe(400);
     });
@@ -707,24 +661,16 @@ describe("password-flow", () => {
     it("should not allow password of a different user to be used", async () => {
       const env = await getEnv();
       const client = testClient(tsoaApp, env);
+      const loginClient = testClient(loginApp, env);
 
-      const typesDoNotWorkWithThisSetup___PARAMS = {
+      const signupResponse = await loginClient.dbconnections.signup.$post({
         json: {
           client_id: "clientId",
           connection: "Username-Password-Authentication",
           email: "new-username-password-user@example.com",
           password: "Password1234!",
         },
-      };
-
-      const signupResponse = await client.dbconnections.signup.$post(
-        typesDoNotWorkWithThisSetup___PARAMS,
-        {
-          headers: {
-            "content-type": "application/json",
-          },
-        },
-      );
+      });
       expect(signupResponse.status).toBe(200);
 
       const loginResponse = await client.co.authenticate.$post(
@@ -827,6 +773,7 @@ describe("password-flow", () => {
     it("should send password reset email for existing user, and allow password to be changed", async () => {
       const env = await getEnv();
       const client = testClient(tsoaApp, env);
+      const loginClient = testClient(loginApp, env);
 
       // foo@example.com is an existing username-password user
       // with password - Test!
@@ -836,20 +783,13 @@ describe("password-flow", () => {
       //-------------------
 
       const passwordResetSendResponse =
-        await client.dbconnections.change_password.$post(
-          {
-            json: {
-              client_id: "clientId",
-              email: "foo@example.com",
-              connection: "Username-Password-Authentication",
-            },
+        await loginClient.dbconnections.change_password.$post({
+          json: {
+            client_id: "clientId",
+            email: "foo@example.com",
+            connection: "Username-Password-Authentication",
           },
-          {
-            headers: {
-              "content-type": "application/json",
-            },
-          },
-        );
+        });
       expect(passwordResetSendResponse.status).toBe(200);
       expect(await passwordResetSendResponse.text()).toBe(
         "We've just sent you an email to reset your password.",
@@ -943,6 +883,7 @@ describe("password-flow", () => {
     it("should reject weak passwords", async () => {
       const env = await getEnv();
       const client = testClient(tsoaApp, env);
+      const loginClient = testClient(loginApp, env);
 
       // foo@example.com is an existing username-password user
       // with password - Test!
@@ -950,7 +891,7 @@ describe("password-flow", () => {
       //-------------------
       // get code to call password reset endpoint
       //-------------------
-      await client.dbconnections.change_password.$post(
+      await loginClient.dbconnections.change_password.$post(
         {
           json: {
             client_id: "clientId",
@@ -969,25 +910,25 @@ describe("password-flow", () => {
       //-------------------
       // reject when try to set weak password
       //-------------------
-      const anyClient = client as any;
-
-      const resetPassword = await anyClient.u["reset-password"].$post({
-        json: {
+      const resetPassword = await loginClient.u["reset-password"].$post({
+        form: {
           // we have unit tests for the util function we use so just doing one unhappy path
           password: "weak-password",
           "re-enter-password": "weak-password",
         },
         query: {
           state,
-          code,
+          // code,
         },
       });
 
       expect(resetPassword.status).toBe(400);
     });
+
     it("should reject non-matching confirmation password", async () => {
       const env = await getEnv();
       const client = testClient(tsoaApp, env);
+      const loginClient = testClient(loginApp, env);
 
       // foo@example.com is an existing username-password user
       // with password - Test!
@@ -995,7 +936,7 @@ describe("password-flow", () => {
       //-------------------
       // get code to call password reset endpoint
       //-------------------
-      await client.dbconnections.change_password.$post(
+      await loginClient.dbconnections.change_password.$post(
         {
           json: {
             client_id: "clientId",
@@ -1033,30 +974,23 @@ describe("password-flow", () => {
     it("should send password reset email for new unvalidated signup AND set email_verified to true", async () => {
       const env = await getEnv();
       const client = testClient(tsoaApp, env);
+      const loginClient = testClient(loginApp, env);
 
-      const typesDoNotWorkWithThisSetup___PARAMS = {
+      const createUserResponse = await loginClient.dbconnections.signup.$post({
         json: {
           client_id: "clientId",
           connection: "Username-Password-Authentication",
           email: "reset-new-user@example.com",
           password: "Password1234!",
         },
-      };
-      const createUserResponse = await client.dbconnections.signup.$post(
-        typesDoNotWorkWithThisSetup___PARAMS,
-        {
-          headers: {
-            "content-type": "application/json",
-          },
-        },
-      );
+      });
       expect(createUserResponse.status).toBe(200);
 
       //-------------------
       // send password reset email even though have never logged in
       //-------------------
       const passwordResetSendResponse =
-        await client.dbconnections.change_password.$post(
+        await loginClient.dbconnections.change_password.$post(
           {
             json: {
               client_id: "clientId",
